# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SINJAPAN logistics AI platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # Shipments
...
  /carriers:
    get:
      operationId: listCarriers
      tags: [carriers]
      summary: List carriers
      responses:
        "200":
          description: List of carriers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Carrier"
    post:
      operationId: createCarrier
      tags: [carriers]
      summary: Create carrier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarrierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Carrier"
  /carriers/{id}:
    get:
      operationId: getCarrier
      tags: [carriers]
      summary: Get carrier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Carrier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Carrier"
    patch:
      operationId: updateCarrier
      tags: [carriers]
      summary: Update carrier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarrierUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Carrier"
    delete:
      operationId: deleteCarrier
      tags: [carriers]
      summary: Delete carrier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # Payments
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        companyName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        companyName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        billingAddress:
          type: ["string", "null"]
        createdAt:
          type: string

    # Shipments
...
    Carrier:
      type: object
      required: [id, companyName, createdAt]
      properties:
        id:
          type: integer
        companyName:
          type: string
        contactName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        serviceAreas:
          type: ["string", "null"]
        vehicleTypes:
          type: ["string", "null"]
        specialties:
          type: ["string", "null"]
        averageCost:
          type: ["number", "null"]
        onTimeRate:
          type: ["number", "null"]
        rating:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        totalOrders:
          type: integer
        createdAt:
          type: string
    CarrierInput:
      type: object
      required: [companyName]
      properties:
        companyName:
          type: string
        contactName:
          type: string
        phone:
          type: string
        email:
          type: string
        serviceAreas:
          type: string
        vehicleTypes:
          type: string
        specialties:
          type: string
        averageCost:
          type: number
        onTimeRate:
          type: number
        rating:
          type: number
        notes:
          type: string
    CarrierUpdate:
      type: object
      properties:
        companyName:
          type: string
        contactName:
          type: string
        phone:
          type: string
        email:
          type: string
        serviceAreas:
          type: string
        vehicleTypes:
          type: string
        specialties:
          type: string
        averageCost:
          type: number
        onTimeRate:
          type: number
        rating:
          type: number
        notes:
          type: string

    # Payments
...